MESSAGES FINAL CHECK

+++ src/DAO/mongodb/MessageMongo.dao.js
@@ -23,7 +23,6 @@ export default class MessageDAO {
         try {
             const result = await messageModel.create(message);
             response.status = "success";
-            response.result = result;
         } catch (error) {
             response.status = "error";
             response.message = "Error al crear el mensaje - DAO: " + error.message;
@@ -51,18 +50,29 @@ export default class MessageDAO {
     };
 
     // Borrar un mensaje - DAO:
-    async deleteMessage(mid) {
+    async deleteMessage(mid, uid) {
         let response = {};
         try {
-            let result = await messageModel.deleteOne({
+            let getSms = await messageModel.findOne({
                 _id: mid
             });
-            if (result.deletedCount === 0) {
-                response.status = "not found message";
-            } else if (result.deletedCount === 1) {
-                response.status = "success";
-                response.result = result;
-            };
+            // Extraemos el uid de la persona que envio el mensaje:
+            let UID = getSms.userId;
+            // Comparamos si el uid de quien intenta borrar el mensaje, conincide con el uid de quien lo envio:
+            if (uid === UID) {
+                // Si la persona solo quiere borrar sus propios mensajes, se le permite la acción:
+                let result = await messageModel.deleteOne({
+                    _id: mid
+                });
+                if (result.deletedCount === 0) {
+                    response.status = "not found message";
+                } else if (result.deletedCount === 1) {
+                    response.status = "success";
+                };
+            } else {
+                // Si la persona intenta borrar un mensaje que no le pertenece, se le deniega la acción:
+                response.status = "unauthorized";
+            }
         } catch (error) {
             response.status = "error";
             response.message = "Error al eliminar el mensaje - DAO: " + error.message;
